# openapi.yaml
openapi: 3.0.3
info:
  title: caceasy API
  version: 0.1.0
  description: API for coupon scanning and wallet management
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-key
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - phone
    Batch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        sku:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - sku
...
    Payout:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: integer
        status:
          type: string
          enum: [pending, completed, failed]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - amount
        - status
...
    OTPRequest:
      type: object
      properties:
        phone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
      required:
        - phone
    OTPVerify:
      type: object
      properties:
        phone:
          type: string
        otp:
          type: string
          pattern: '^\d{6}$'
      required:
        - phone
        - otp
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - user
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
paths:
...
  /auth/otp:
    post:
      summary: Request OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPRequest'
      responses:
        '202':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify:
    post:
      summary: Verify OTP and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerify'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /batches:
    post:
      summary: Create a new batch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                sku:
                  type: string
              required:
                - name
                - sku
      responses:
        '201':
          description: Batch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/payouts:
    get:
      summary: Get payouts (admin only)
      security:
        - adminApiKey: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed]
      responses:
        '200':
          description: List of payouts
          content:
            application/json:
              schema:
                type: object
                properties:
                  payouts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Payout'
        '401':
          description: Unauthorized
  /admin/flagged:
    get:
      summary: Get flagged activities (admin only)
      security:
        - adminApiKey: []
      responses:
        '200':
          description: List of flagged activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  flagged:
                    type: array
                    items:
                      type: object
        '401':
          description: Unauthorized